Add all ore names to config for reference
When generating the config, OreDupeFix will now add all of the
registered ore dictionary names to PreferredOres. This is so you
can easily fill in the mod names yourself for unrecognized ores,
instead of having to check the console log and add new entries
for the appropriate names. Just fill in the blanks.

Example output: https://gist.github.com/agaricusb/5125787
Thanks to narc for this suggestion:
http://forum.industrial-craft.net/index.php?page=Thread&postID=99750#post99750